% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPimportance.R
\name{PPimportance}
\alias{PPimportance}
\title{predict PPtree}
\usage{
PPimportance(PPtreeregOBJ,...)
}
\arguments{
\item{PPtreeregOBJ}{PPtreereg object}

\item{...}{arguments to be passed to methods}
}
\value{
imp_node_split predicted values

imp_node_final MSE of the predicted values

imp_var MSE of the predicted values
}
\description{
Predict class for the test set and calculate prediction error 
After finding tree structure, predict class for the test set and calculate 
prediction error.
}
\examples{
data(mtcars)
n <- nrow(mtcars)
tot <- c(1:n)
n.train <- round(n*0.8)
train <- sample(tot,n.train)
test <- tot[-train]
Tree.result <- PPTreereg(mpg~.,mtcars[train,],
                           final.rule=1,DEPTH=2)
PPimportance(Tree.result)
}
\keyword{tree}
